Сигнал и вспомогательный код перенесен в schooling/utils.py. Исправлены круговые импорты. Отправка сообщений теперь формируется через asyncio.gather.
