+++ src/schooling/apps.py
@@ -11,4 +11,3 @@ class SchoolingConfig(AppConfig):
     def ready(self) -> None:
         """Подлючает сигналы."""
         from schooling import utils # noqa
-        from schooling import signals # noqa
