+++ src/schooling/utils.py
@@ -1,15 +1,81 @@
 import asyncio
 from datetime import timedelta
 
+from telegram import InlineKeyboardButton, InlineKeyboardMarkup
+from telegram.ext import CallbackContext
+from telegram.error import BadRequest
+from django.utils import timezone
 from asgiref.sync import async_to_sync, sync_to_async
 from django.conf import settings
 from django.db.models.signals import post_save
 from django.dispatch import receiver
-from telegram import Bot
-from telegram.error import BadRequest
 
 from bot.keyboards import get_root_markup
 from schooling.models import Student, Teacher, Lesson
+
+
+async def send_message_async(
+    bot, chat_id, message_text, reply_markup
+):
+    try:
+        await bot.send_message(
+            chat_id=chat_id,
+            text=message_text,
+            reply_markup=reply_markup,
+        )
+    except BadRequest:
+        print(f'Чат с id {chat_id} не найден!')
+
+
+async def gather_send_messages_to_users(
+    bot, chat_ids, message_text, reply_markup
+):
+    tasks = [
+        send_message_async(
+            bot, chat_id, message_text, reply_markup
+        ) for chat_id in chat_ids
+    ]
+    await asyncio.gather(*tasks)
+
+
+async def send_lesson_end_notification(context: CallbackContext):
+    from telegram import Bot as TGBot
+    bot = TGBot(token=settings.TELEGRAM_TOKEN)
+
+    teacher_chat_id = context.job.data.get('teacher_chat_id')
+    student_chat_id = context.job.data.get('student_chat_id')
+    lesson_id = context.job.data.get('lesson_id')
+
+    keyboard = [[
+        InlineKeyboardButton('Да', callback_data=f'yes {lesson_id}'),
+        InlineKeyboardButton('Нет', callback_data=f'no {lesson_id}'),
+    ]]
+    reply_markup = InlineKeyboardMarkup(keyboard)
+
+    message_text = 'Было ли занятие?'
+
+    chat_ids = (teacher_chat_id, student_chat_id)
+    await gather_send_messages_to_users(
+        bot, chat_ids, message_text, reply_markup
+    )
+
+
+@receiver(post_save, sender=Lesson)
+async def schedule_lesson_end_notification(sender, instance, **kwargs):
+    from bot.bot_interface import Bot
+    bot = Bot()
+    lesson_end_time = instance.datetime_end
+    if lesson_end_time > timezone.now():
+        bot._app.job_queue.run_once(
+            send_lesson_end_notification,
+            when=lesson_end_time,
+            name=f'lesson_end_{instance.id}',
+            data={
+                'teacher_chat_id': instance.teacher_id.telegram_id,
+                'student_chat_id': instance.student_id.telegram_id,
+                'lesson_id': instance.id,
+            },
+        )
 
 
 @async_to_sync
@@ -17,7 +83,8 @@ async def send_message_to_user(bot_token, user_id,
                                message_text,
                                reply_markup=None):
     """Инициативно отправляет сообщение."""
-    bot = Bot(token=bot_token)
+    from telegram import Bot as TGBot
+    bot = TGBot(token=bot_token)
     try:
         await bot.send_message(chat_id=user_id,
                                text=message_text,
